+++ common/persistence/nosql/nosqlplugin/cassandra/workflow.go
@@ -23,646 +23,10 @@ package cassandra
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/uber/cadence/common/persistence"
 	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin"
 	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql"
-)
-
-const (
-	templateWorkflowExecutionType = `{` +
-		`domain_id: ?, ` +
-		`workflow_id: ?, ` +
-		`run_id: ?, ` +
-		`parent_domain_id: ?, ` +
-		`parent_workflow_id: ?, ` +
-		`parent_run_id: ?, ` +
-		`initiated_id: ?, ` +
-		`completion_event_batch_id: ?, ` +
-		`completion_event: ?, ` +
-		`completion_event_data_encoding: ?, ` +
-		`task_list: ?, ` +
-		`workflow_type_name: ?, ` +
-		`workflow_timeout: ?, ` +
-		`decision_task_timeout: ?, ` +
-		`execution_context: ?, ` +
-		`state: ?, ` +
-		`close_status: ?, ` +
-		`last_first_event_id: ?, ` +
-		`last_event_task_id: ?, ` +
-		`next_event_id: ?, ` +
-		`last_processed_event: ?, ` +
-		`start_time: ?, ` +
-		`last_updated_time: ?, ` +
-		`create_request_id: ?, ` +
-		`signal_count: ?, ` +
-		`history_size: ?, ` +
-		`decision_version: ?, ` +
-		`decision_schedule_id: ?, ` +
-		`decision_started_id: ?, ` +
-		`decision_request_id: ?, ` +
-		`decision_timeout: ?, ` +
-		`decision_attempt: ?, ` +
-		`decision_timestamp: ?, ` +
-		`decision_scheduled_timestamp: ?, ` +
-		`decision_original_scheduled_timestamp: ?, ` +
-		`cancel_requested: ?, ` +
-		`cancel_request_id: ?, ` +
-		`sticky_task_list: ?, ` +
-		`sticky_schedule_to_start_timeout: ?,` +
-		`client_library_version: ?, ` +
-		`client_feature_version: ?, ` +
-		`client_impl: ?, ` +
-		`auto_reset_points: ?, ` +
-		`auto_reset_points_encoding: ?, ` +
-		`attempt: ?, ` +
-		`has_retry_policy: ?, ` +
-		`init_interval: ?, ` +
-		`backoff_coefficient: ?, ` +
-		`max_interval: ?, ` +
-		`expiration_time: ?, ` +
-		`max_attempts: ?, ` +
-		`non_retriable_errors: ?, ` +
-		`event_store_version: ?, ` +
-		`branch_token: ?, ` +
-		`cron_schedule: ?, ` +
-		`expiration_seconds: ?, ` +
-		`search_attributes: ?, ` +
-		`memo: ? ` +
-		`}`
-
-	templateTransferTaskType = `{` +
-		`domain_id: ?, ` +
-		`workflow_id: ?, ` +
-		`run_id: ?, ` +
-		`visibility_ts: ?, ` +
-		`task_id: ?, ` +
-		`target_domain_id: ?, ` +
-		`target_workflow_id: ?, ` +
-		`target_run_id: ?, ` +
-		`target_child_workflow_only: ?, ` +
-		`task_list: ?, ` +
-		`type: ?, ` +
-		`schedule_id: ?, ` +
-		`record_visibility: ?, ` +
-		`version: ?` +
-		`}`
-
-	templateCrossClusterTaskType = templateTransferTaskType
-
-	templateReplicationTaskType = `{` +
-		`domain_id: ?, ` +
-		`workflow_id: ?, ` +
-		`run_id: ?, ` +
-		`task_id: ?, ` +
-		`type: ?, ` +
-		`first_event_id: ?,` +
-		`next_event_id: ?,` +
-		`version: ?,` +
-		`scheduled_id: ?, ` +
-		`event_store_version: ?, ` +
-		`branch_token: ?, ` +
-		`new_run_event_store_version: ?, ` +
-		`new_run_branch_token: ?, ` +
-		`created_time: ? ` +
-		`}`
-
-	templateTimerTaskType = `{` +
-		`domain_id: ?, ` +
-		`workflow_id: ?, ` +
-		`run_id: ?, ` +
-		`visibility_ts: ?, ` +
-		`task_id: ?, ` +
-		`type: ?, ` +
-		`timeout_type: ?, ` +
-		`event_id: ?, ` +
-		`schedule_attempt: ?, ` +
-		`version: ?` +
-		`}`
-
-	templateActivityInfoType = `{` +
-		`version: ?,` +
-		`schedule_id: ?, ` +
-		`scheduled_event_batch_id: ?, ` +
-		`scheduled_event: ?, ` +
-		`scheduled_time: ?, ` +
-		`started_id: ?, ` +
-		`started_event: ?, ` +
-		`started_time: ?, ` +
-		`activity_id: ?, ` +
-		`request_id: ?, ` +
-		`details: ?, ` +
-		`schedule_to_start_timeout: ?, ` +
-		`schedule_to_close_timeout: ?, ` +
-		`start_to_close_timeout: ?, ` +
-		`heart_beat_timeout: ?, ` +
-		`cancel_requested: ?, ` +
-		`cancel_request_id: ?, ` +
-		`last_hb_updated_time: ?, ` +
-		`timer_task_status: ?, ` +
-		`attempt: ?, ` +
-		`task_list: ?, ` +
-		`started_identity: ?, ` +
-		`has_retry_policy: ?, ` +
-		`init_interval: ?, ` +
-		`backoff_coefficient: ?, ` +
-		`max_interval: ?, ` +
-		`expiration_time: ?, ` +
-		`max_attempts: ?, ` +
-		`non_retriable_errors: ?, ` +
-		`last_failure_reason: ?, ` +
-		`last_worker_identity: ?, ` +
-		`last_failure_details: ?, ` +
-		`event_data_encoding: ?` +
-		`}`
-
-	templateTimerInfoType = `{` +
-		`version: ?,` +
-		`timer_id: ?, ` +
-		`started_id: ?, ` +
-		`expiry_time: ?, ` +
-		`task_id: ?` +
-		`}`
-
-	templateChildExecutionInfoType = `{` +
-		`version: ?,` +
-		`initiated_id: ?, ` +
-		`initiated_event_batch_id: ?, ` +
-		`initiated_event: ?, ` +
-		`started_id: ?, ` +
-		`started_workflow_id: ?, ` +
-		`started_run_id: ?, ` +
-		`started_event: ?, ` +
-		`create_request_id: ?, ` +
-		`event_data_encoding: ?, ` +
-		`domain_name: ?, ` +
-		`workflow_type_name: ?, ` +
-		`parent_close_policy: ?` +
-		`}`
-
-	templateRequestCancelInfoType = `{` +
-		`version: ?,` +
-		`initiated_id: ?, ` +
-		`initiated_event_batch_id: ?, ` +
-		`cancel_request_id: ? ` +
-		`}`
-
-	templateSignalInfoType = `{` +
-		`version: ?,` +
-		`initiated_id: ?, ` +
-		`initiated_event_batch_id: ?, ` +
-		`signal_request_id: ?, ` +
-		`signal_name: ?, ` +
-		`input: ?, ` +
-		`control: ?` +
-		`}`
-
-	templateChecksumType = `{` +
-		`version: ?, ` +
-		`flavor: ?, ` +
-		`value: ? ` +
-		`}`
-
-	templateUpdateCurrentWorkflowExecutionQuery = `UPDATE executions USING TTL 0 ` +
-		`SET current_run_id = ?,
-execution = {run_id: ?, create_request_id: ?, state: ?, close_status: ?},
-workflow_last_write_version = ?,
-workflow_state = ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? ` +
-		`IF current_run_id = ? `
-
-	templateUpdateCurrentWorkflowExecutionForNewQuery = templateUpdateCurrentWorkflowExecutionQuery +
-		`and workflow_last_write_version = ? ` +
-		`and workflow_state = ? `
-
-	templateCreateCurrentWorkflowExecutionQuery = `INSERT INTO executions (` +
-		`shard_id, type, domain_id, workflow_id, run_id, visibility_ts, task_id, current_run_id, execution, workflow_last_write_version, workflow_state) ` +
-		`VALUES(?, ?, ?, ?, ?, ?, ?, ?, {run_id: ?, create_request_id: ?, state: ?, close_status: ?}, ?, ?) IF NOT EXISTS USING TTL 0 `
-
-	templateCreateWorkflowExecutionWithVersionHistoriesQuery = `INSERT INTO executions (` +
-		`shard_id, domain_id, workflow_id, run_id, type, execution, next_event_id, visibility_ts, task_id, version_histories, version_histories_encoding, checksum, workflow_last_write_version, workflow_state) ` +
-		`VALUES(?, ?, ?, ?, ?, ` + templateWorkflowExecutionType + `, ?, ?, ?, ?, ?, ` + templateChecksumType + `, ?, ?) IF NOT EXISTS `
-
-	templateCreateTransferTaskQuery = `INSERT INTO executions (` +
-		`shard_id, type, domain_id, workflow_id, run_id, transfer, visibility_ts, task_id) ` +
-		`VALUES(?, ?, ?, ?, ?, ` + templateTransferTaskType + `, ?, ?)`
-
-	templateCreateCrossClusterTaskQuery = `INSERT INTO executions (` +
-		`shard_id, type, domain_id, workflow_id, run_id, cross_cluster, visibility_ts, task_id) ` +
-		`VALUES(?, ?, ?, ?, ?, ` + templateCrossClusterTaskType + `, ?, ?)`
-
-	templateCreateReplicationTaskQuery = `INSERT INTO executions (` +
-		`shard_id, type, domain_id, workflow_id, run_id, replication, visibility_ts, task_id) ` +
-		`VALUES(?, ?, ?, ?, ?, ` + templateReplicationTaskType + `, ?, ?)`
-
-	templateCreateTimerTaskQuery = `INSERT INTO executions (` +
-		`shard_id, type, domain_id, workflow_id, run_id, timer, visibility_ts, task_id) ` +
-		`VALUES(?, ?, ?, ?, ?, ` + templateTimerTaskType + `, ?, ?)`
-
-	templateUpdateLeaseQuery = `UPDATE executions ` +
-		`SET range_id = ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? ` +
-		`IF range_id = ?`
-
-	// TODO: remove replication_state after all 2DC workflows complete
-	templateGetWorkflowExecutionQuery = `SELECT execution, replication_state, activity_map, timer_map, ` +
-		`child_executions_map, request_cancel_map, signal_map, signal_requested, buffered_events_list, ` +
-		`buffered_replication_tasks_map, version_histories, version_histories_encoding, checksum ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ?`
-
-	templateGetCurrentExecutionQuery = `SELECT current_run_id, execution, workflow_last_write_version ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ?`
-
-	templateListCurrentExecutionsQuery = `SELECT domain_id, workflow_id, run_id, current_run_id, workflow_state ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ?`
-
-	templateIsWorkflowExecutionExistsQuery = `SELECT shard_id, type, domain_id, workflow_id, run_id, visibility_ts, task_id ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ?`
-
-	templateListWorkflowExecutionQuery = `SELECT run_id, execution, version_histories, version_histories_encoding ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ?`
-
-	templateCheckWorkflowExecutionQuery = `UPDATE executions ` +
-		`SET next_event_id = ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? ` +
-		`IF next_event_id = ?`
-
-	templateUpdateWorkflowExecutionWithVersionHistoriesQuery = `UPDATE executions ` +
-		`SET execution = ` + templateWorkflowExecutionType +
-		`, next_event_id = ? ` +
-		`, version_histories = ? ` +
-		`, version_histories_encoding = ? ` +
-		`, checksum = ` + templateChecksumType +
-		`, workflow_last_write_version = ? ` +
-		`, workflow_state = ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? ` +
-		`IF next_event_id = ? `
-
-	templateUpdateActivityInfoQuery = `UPDATE executions ` +
-		`SET activity_map[ ? ] =` + templateActivityInfoType + ` ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetActivityInfoQuery = `UPDATE executions ` +
-		`SET activity_map = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateUpdateTimerInfoQuery = `UPDATE executions ` +
-		`SET timer_map[ ? ] =` + templateTimerInfoType + ` ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetTimerInfoQuery = `UPDATE executions ` +
-		`SET timer_map = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateUpdateChildExecutionInfoQuery = `UPDATE executions ` +
-		`SET child_executions_map[ ? ] =` + templateChildExecutionInfoType + ` ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetChildExecutionInfoQuery = `UPDATE executions ` +
-		`SET child_executions_map = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateUpdateRequestCancelInfoQuery = `UPDATE executions ` +
-		`SET request_cancel_map[ ? ] =` + templateRequestCancelInfoType + ` ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetRequestCancelInfoQuery = `UPDATE executions ` +
-		`SET request_cancel_map = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateUpdateSignalInfoQuery = `UPDATE executions ` +
-		`SET signal_map[ ? ] =` + templateSignalInfoType + ` ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetSignalInfoQuery = `UPDATE executions ` +
-		`SET signal_map = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateUpdateSignalRequestedQuery = `UPDATE executions ` +
-		`SET signal_requested = signal_requested + ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateResetSignalRequestedQuery = `UPDATE executions ` +
-		`SET signal_requested = ?` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateAppendBufferedEventsQuery = `UPDATE executions ` +
-		`SET buffered_events_list = buffered_events_list + ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteBufferedEventsQuery = `UPDATE executions ` +
-		`SET buffered_events_list = [] ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteActivityInfoQuery = `DELETE activity_map[ ? ] ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteTimerInfoQuery = `DELETE timer_map[ ? ] ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteChildExecutionInfoQuery = `DELETE child_executions_map[ ? ] ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteRequestCancelInfoQuery = `DELETE request_cancel_map[ ? ] ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteSignalInfoQuery = `DELETE signal_map[ ? ] ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteWorkflowExecutionMutableStateQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateDeleteWorkflowExecutionCurrentRowQuery = templateDeleteWorkflowExecutionMutableStateQuery + " if current_run_id = ? "
-
-	templateDeleteWorkflowExecutionSignalRequestedQuery = `UPDATE executions ` +
-		`SET signal_requested = signal_requested - ? ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ? `
-
-	templateGetTransferTasksQuery = `SELECT transfer ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id > ? ` +
-		`and task_id <= ?`
-
-	templateGetCrossClusterTasksQuery = `SELECT cross_cluster ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id > ? ` +
-		`and task_id <= ?`
-
-	templateGetReplicationTasksQuery = `SELECT replication ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id > ? ` +
-		`and task_id <= ?`
-
-	templateGetDLQSizeQuery = `SELECT count(1) as count ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ?`
-
-	templateCompleteTransferTaskQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id = ?`
-
-	templateRangeCompleteTransferTaskQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id > ? ` +
-		`and task_id <= ?`
-
-	templateCompleteCrossClusterTaskQuery = templateCompleteTransferTaskQuery
-
-	templateRangeCompleteCrossClusterTaskQuery = templateRangeCompleteTransferTaskQuery
-
-	templateCompleteReplicationTaskBeforeQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ? ` +
-		`and run_id = ? ` +
-		`and visibility_ts = ? ` +
-		`and task_id <= ?`
-
-	templateCompleteReplicationTaskQuery = templateCompleteTransferTaskQuery
-
-	templateRangeCompleteReplicationTaskQuery = templateRangeCompleteTransferTaskQuery
-
-	templateGetTimerTasksQuery = `SELECT timer ` +
-		`FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ?` +
-		`and domain_id = ? ` +
-		`and workflow_id = ?` +
-		`and run_id = ?` +
-		`and visibility_ts >= ? ` +
-		`and visibility_ts < ?`
-
-	templateCompleteTimerTaskQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ?` +
-		`and run_id = ?` +
-		`and visibility_ts = ? ` +
-		`and task_id = ?`
-
-	templateRangeCompleteTimerTaskQuery = `DELETE FROM executions ` +
-		`WHERE shard_id = ? ` +
-		`and type = ? ` +
-		`and domain_id = ? ` +
-		`and workflow_id = ?` +
-		`and run_id = ?` +
-		`and visibility_ts >= ? ` +
-		`and visibility_ts < ?`
 )
 
 func (db *cdb) InsertWorkflowExecutionWithTasks(
@@ -680,21 +44,24 @@ func (db *cdb) InsertWorkflowExecutionWithTasks(
 	signalInfoMap map[int64]*persistence.SignalInfo,
 	signalRequestedIDs []string,
 	shardCondition *nosqlplugin.ShardCondition,
-) (*nosqlplugin.ConditionFailureDetails, error) {
+) (*nosqlplugin.ConditionFailureReason, error) {
 	shardID := shardCondition.ShardID
 	domainID := execution.DomainID
 	workflowID := execution.WorkflowID
 	runID := execution.RunID
 
 	batch := db.session.NewBatch(gocql.LoggedBatch).WithContext(ctx)
+
 	err := db.createOrUpdateCurrentWorkflow(batch, shardID, domainID, workflowID, currentWorkflowRequest)
 	if err != nil {
 		return nil, err
 	}
+
 	err = db.createWorkflowExecution(batch, shardID, domainID, workflowID, runID, execution)
 	if err != nil {
 		return nil, err
 	}
+
 	err = db.updateActivityInfos(batch, shardID, domainID, workflowID, runID, activityInfoMap, nil)
 	if err != nil {
 		return nil, err
@@ -720,553 +87,27 @@ func (db *cdb) InsertWorkflowExecutionWithTasks(
 		return nil, err
 	}
 
-	return nil, nil
-}
-
-func (db *cdb) createReplicationTasks(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	transferTasks []*nosqlplugin.ReplicationTask,
-) error {
-	for _, task := range transferTasks {
-		batch.Query(templateCreateReplicationTaskQuery,
-			shardID,
-			rowTypeReplicationTask,
-			rowTypeReplicationDomainID,
-			rowTypeReplicationWorkflowID,
-			rowTypeReplicationRunID,
-			domainID,
-			workflowID,
-			runID,
-			task.TaskID,
-			task.Type,
-			task.FirstEventID,
-			task.NextEventID,
-			task.Version,
-			activityScheduleID,
-			p.EventStoreVersion,
-			branchToken,
-			p.EventStoreVersion,
-			newRunBranchToken,
-			task.GetVisibilityTimestamp().UnixNano(),
-			defaultVisibilityTimestamp,
-			task.GetTaskID())
+	err = db.createTransferTasks(batch, shardID, domainID, workflowID, runID, transferTasks)
+	if err != nil {
+		return nil, err
 	}
-	return nil
-}
-
-func (db *cdb) createTransferTasks(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	transferTasks []*nosqlplugin.TransferTask,
-) error {
-	for _, task := range transferTasks {
-		batch.Query(templateCreateTransferTaskQuery,
-			shardID,
-			rowTypeTransferTask,
-			rowTypeTransferDomainID,
-			rowTypeTransferWorkflowID,
-			rowTypeTransferRunID,
-			domainID,
-			workflowID,
-			runID,
-			task.VisibilityTimestamp,
-			task.TaskID,
-			task.TargetDomainID,
-			task.TargetWorkflowID,
-			task.TargetRunID,
-			task.TargetChildWorkflowOnly,
-			task.TaskList,
-			task.Type,
-			task.ScheduleID,
-			task.RecordVisibility,
-			task.Version,
-			defaultVisibilityTimestamp, // ?? TODO?? should we use task.VisibilityTimestamp?
-			task.TaskID)
+	err = db.createReplicationTasks(batch, shardID, domainID, workflowID, runID, replicationTasks)
+	if err != nil {
+		return nil, err
 	}
-	return nil
-}
-
-func (db *cdb) createCrossClusterTasks(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	transferTasks []*nosqlplugin.CrossClusterTask,
-) error {
-	for _, task := range transferTasks {
-		batch.Query(templateCreateCrossClusterTaskQuery,
-			shardID,
-			rowTypeCrossClusterTask,
-			rowTypeCrossClusterDomainID,
-			task.TargetCluster,
-			rowTypeCrossClusterRunID,
-			domainID,
-			workflowID,
-			runID,
-			task.VisibilityTimestamp,
-			task.TaskID,
-			task.TargetDomainID,
-			task.TargetWorkflowID,
-			task.TargetRunID,
-			task.TargetChildWorkflowOnly,
-			task.TaskList,
-			task.Type,
-			task.ScheduleID,
-			task.RecordVisibility,
-			task.Version,
-			defaultVisibilityTimestamp, // ?? TODO?? should we use task.VisibilityTimestamp?
-			task.TaskID,
-		)
+	err = db.createCrossClusterTasks(batch, shardID, domainID, workflowID, runID, crossClusterTasks)
+	if err != nil {
+		return nil, err
 	}
-	return nil
-}
-
-func (db *cdb) updateSignalsRequested(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	signalReqIDs []string,
-	deleteSignalReqIDs []string,
-) error {
-
-	if len(signalReqIDs) > 0 {
-		batch.Query(templateUpdateSignalRequestedQuery,
-			signalReqIDs,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
+	err = db.createTimerTasks(batch, shardID, domainID, workflowID, runID, timerTasks)
+	if err != nil {
+		return nil, err
 	}
 
-	if len(deleteSignalReqIDs) > 0 {
-		batch.Query(templateDeleteWorkflowExecutionSignalRequestedQuery,
-			deleteSignalReqIDs,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) updateSignalInfos(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	signalInfos map[int64]*persistence.SignalInfo,
-	deleteInfos []int64,
-) error {
-	for _, c := range signalInfos {
-		batch.Query(templateUpdateSignalInfoQuery,
-			c.InitiatedID,
-			c.Version,
-			c.InitiatedID,
-			c.InitiatedEventBatchID,
-			c.SignalRequestID,
-			c.SignalName,
-			c.Input,
-			c.Control,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
+	err = db.assertShardRangeID(batch, shardID, shardCondition.RangeID)
+	if err != nil {
+		return nil, err
 	}
 
-	// deleteInfos are the initiatedIDs for SignalInfo being deleted
-	for _, deleteInfo := range deleteInfos {
-		batch.Query(templateDeleteSignalInfoQuery,
-			deleteInfo,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) updateRequestCancelInfos(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	requestCancelInfos map[int64]*persistence.RequestCancelInfo,
-	deleteInfos []int64,
-) error {
-
-	for _, c := range requestCancelInfos {
-		batch.Query(templateUpdateRequestCancelInfoQuery,
-			c.InitiatedID,
-			c.Version,
-			c.InitiatedID,
-			c.InitiatedEventBatchID,
-			c.CancelRequestID,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-
-	// deleteInfos are the initiatedIDs for RequestCancelInfo being deleted
-	for _, deleteInfo := range deleteInfos {
-		batch.Query(templateDeleteRequestCancelInfoQuery,
-			deleteInfo,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) updateChildExecutionInfos(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	childExecutionInfos map[int64]*persistence.InternalChildExecutionInfo,
-	deleteInfos []int64,
-) error {
-
-	for _, c := range childExecutionInfos {
-		startedRunID := emptyRunID
-		if c.StartedRunID != "" {
-			startedRunID = c.StartedRunID
-		}
-
-		batch.Query(templateUpdateChildExecutionInfoQuery,
-			c.InitiatedID,
-			c.Version,
-			c.InitiatedID,
-			c.InitiatedEventBatchID,
-			c.InitiatedEvent.Data,
-			c.StartedID,
-			c.StartedWorkflowID,
-			startedRunID,
-			c.StartedEvent.Data,
-			c.CreateRequestID,
-			c.InitiatedEvent.GetEncodingString(),
-			c.DomainName,
-			c.WorkflowTypeName,
-			int32(c.ParentClosePolicy),
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-
-	// deleteInfos are the initiatedIDs for ChildInfo being deleted
-	for _, deleteInfo := range deleteInfos {
-		batch.Query(templateDeleteChildExecutionInfoQuery,
-			deleteInfo,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) updateTimerInfos(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	timerInfos map[string]*persistence.TimerInfo,
-	deleteInfos []string,
-) error {
-	for _, timerInfo := range timerInfos {
-		batch.Query(templateUpdateTimerInfoQuery,
-			timerInfo.TimerID,
-			timerInfo.Version,
-			timerInfo.TimerID,
-			timerInfo.StartedID,
-			timerInfo.ExpiryTime,
-			timerInfo.TaskStatus,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-
-	for _, deleteInfo := range deleteInfos {
-		batch.Query(templateDeleteTimerInfoQuery,
-			deleteInfo,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) updateActivityInfos(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	activityInfos map[int64]*persistence.InternalActivityInfo,
-	deleteInfos []int64,
-) error {
-	for _, a := range activityInfos {
-		batch.Query(templateUpdateActivityInfoQuery,
-			a.ScheduleID,
-			a.Version,
-			a.ScheduleID,
-			a.ScheduledEventBatchID,
-			a.ScheduledEvent.Data,
-			a.ScheduledTime,
-			a.StartedID,
-			a.StartedEvent.Data,
-			a.StartedTime,
-			a.ActivityID,
-			a.RequestID,
-			a.Details,
-			int32(a.ScheduleToStartTimeout.Seconds()),
-			int32(a.ScheduleToCloseTimeout.Seconds()),
-			int32(a.StartToCloseTimeout.Seconds()),
-			int32(a.HeartbeatTimeout.Seconds()),
-			a.CancelRequested,
-			a.CancelRequestID,
-			a.LastHeartBeatUpdatedTime,
-			a.TimerTaskStatus,
-			a.Attempt,
-			a.TaskList,
-			a.StartedIdentity,
-			a.HasRetryPolicy,
-			int32(a.InitialInterval.Seconds()),
-			a.BackoffCoefficient,
-			int32(a.MaximumInterval.Seconds()),
-			a.ExpirationTime,
-			a.MaximumAttempts,
-			a.NonRetriableErrors,
-			a.LastFailureReason,
-			a.LastWorkerIdentity,
-			a.LastFailureDetails,
-			a.ScheduledEvent.GetEncodingString(),
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-
-	for _, deleteInfo := range deleteInfos {
-		batch.Query(templateDeleteActivityInfoQuery,
-			deleteInfo,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			runID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID)
-	}
-	return nil
-}
-
-func (db *cdb) createWorkflowExecution(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	runID string,
-	execution *nosqlplugin.WorkflowExecutionRow,
-) error {
-	batch.Query(templateCreateWorkflowExecutionWithVersionHistoriesQuery,
-		shardID,
-		domainID,
-		workflowID,
-		runID,
-		rowTypeExecution,
-		domainID,
-		workflowID,
-		runID,
-		execution.ParentDomainID,
-		execution.ParentWorkflowID,
-		execution.ParentRunID,
-		execution.InitiatedID,
-		execution.CompletionEventBatchID,
-		execution.CompletionEvent.Data,
-		execution.CompletionEvent.GetEncodingString(),
-		execution.TaskList,
-		execution.WorkflowTypeName,
-		int32(execution.WorkflowTimeout.Seconds()),
-		int32(execution.DecisionStartToCloseTimeout.Seconds()),
-		execution.ExecutionContext,
-		execution.State,
-		execution.CloseStatus,
-		execution.LastFirstEventID,
-		execution.LastEventTaskID,
-		execution.NextEventID,
-		execution.LastProcessedEvent,
-		execution.StartTimestamp,
-		execution.LastUpdatedTimestamp,
-		execution.CreateRequestID,
-		execution.SignalCount,
-		execution.HistorySize,
-		execution.DecisionVersion,
-		execution.DecisionScheduleID,
-		execution.DecisionStartedID,
-		execution.DecisionRequestID,
-		int32(execution.DecisionTimeout.Seconds()),
-		execution.DecisionAttempt,
-		execution.DecisionStartedTimestamp.UnixNano(),
-		execution.DecisionScheduledTimestamp.UnixNano(),
-		execution.DecisionOriginalScheduledTimestamp.UnixNano(),
-		execution.CancelRequested,
-		execution.CancelRequestID,
-		execution.StickyTaskList,
-		int32(execution.StickyScheduleToStartTimeout.Seconds()),
-		execution.ClientLibraryVersion,
-		execution.ClientFeatureVersion,
-		execution.ClientImpl,
-		execution.AutoResetPoints.Data,
-		execution.AutoResetPoints.GetEncodingString(),
-		execution.Attempt,
-		execution.HasRetryPolicy,
-		int32(execution.InitialInterval.Seconds()),
-		execution.BackoffCoefficient,
-		int32(execution.MaximumInterval.Seconds()),
-		execution.ExpirationTime,
-		execution.MaximumAttempts,
-		execution.NonRetriableErrors,
-		persistence.EventStoreVersion,
-		execution.BranchToken,
-		execution.CronSchedule,
-		int32(execution.ExpirationSeconds.Seconds()),
-		execution.SearchAttributes,
-		execution.Memo,
-		execution.NextEventID,
-		defaultVisibilityTimestamp,
-		rowTypeExecutionTaskID,
-		execution.VersionHistories.Data,
-		execution.VersionHistories.GetEncodingString(),
-		execution.Checksums.Version,
-		execution.Checksums.Flavor,
-		execution.Checksums.Value,
-		execution.LastWriteVersion,
-		execution.State,
-	)
-	return nil
-}
-
-func (db *cdb) createOrUpdateCurrentWorkflow(
-	batch gocql.Batch,
-	shardID int,
-	domainID string,
-	workflowID string,
-	request *nosqlplugin.CurrentWorkflowWriteRequest,
-) error {
-	switch request.WriteMode {
-	case nosqlplugin.CurrentWorkflowWriteModeNoop:
-		return nil
-	case nosqlplugin.CurrentWorkflowWriteModeInsert:
-		batch.Query(templateCreateCurrentWorkflowExecutionQuery,
-			shardID,
-			rowTypeExecution,
-			domainID,
-			workflowID,
-			permanentRunID,
-			defaultVisibilityTimestamp,
-			rowTypeExecutionTaskID,
-			request.RunID,
-			request.RunID,
-			request.CreateRequestID,
-			request.State,
-			request.CloseStatus,
-			request.LastWriteVersion,
-			request.State,
-		)
-	case nosqlplugin.CurrentWorkflowWriteModeUpdate:
-		if request.PreviousLastWriteVersion != nil && request.PreviousState != nil {
-			batch.Query(templateUpdateCurrentWorkflowExecutionForNewQuery,
-				request.RunID,
-				request.RunID,
-				request.CreateRequestID,
-				request.State,
-				request.CloseStatus,
-				request.LastWriteVersion,
-				request.State,
-				shardID,
-				rowTypeExecution,
-				domainID,
-				workflowID,
-				permanentRunID,
-				defaultVisibilityTimestamp,
-				rowTypeExecutionTaskID,
-				*request.PreviousRunID,
-				*request.PreviousLastWriteVersion,
-				*request.PreviousState,
-			)
-		} else {
-			batch.Query(templateUpdateCurrentWorkflowExecutionQuery,
-				request.RunID,
-				request.RunID,
-				request.CreateRequestID,
-				request.State,
-				request.CloseStatus,
-				request.LastWriteVersion,
-				request.State,
-				shardID,
-				rowTypeExecution,
-				domainID,
-				workflowID,
-				permanentRunID,
-				defaultVisibilityTimestamp,
-				rowTypeExecutionTaskID,
-				*request.PreviousRunID,
-			)
-		}
-	default:
-		return fmt.Errorf("unknown mode %v", request.WriteMode)
-	}
-	return nil
+	return db.executeCreateWorkflowBatchTransaction(batch, currentWorkflowRequest, execution, shardCondition)
 }
